@@ -24,7 +24,7 @@ func run() int {
 	fs := flag.NewFlagSetWithEnvPrefix(os.Args[0], "E3DC", flag.ContinueOnError)
 	fs.String(flag.DefaultConfigFlagname, ".config", "path to config file")
 	fs.StringVar(&conf.host, "host", "", "e3dc server host")
-	fs.UintVar(&conf.port, "port", 5033, "e3dc server host port") //nolint:gomnd
+	fs.UintVar(&conf.port, "port", 5033, "e3dc server host port") //nolint:mnd
 	fs.StringVar(&conf.user, "user", "", "e3dc user")
 	fs.StringVar(&conf.password, "password", "", "e3dc password (consider using a config file or environment variable)")
 	fs.StringVar(&conf.key, "key", "", "rscp key")
@@ -33,7 +33,7 @@ func run() int {
 	}
 	c, err := rscp.NewClient(rscp.ClientConfig{
 		Address:     conf.host,
-		Port:        uint16(conf.port),
+		Port:        uint16(conf.port), //nolint:gosec
 		Username:    conf.user,
 		Password:    conf.password,
 		Key:         conf.key,
