@@ -34,7 +34,7 @@ type ClientConfig struct {
 
 // defaultClientConfig defines the default config values used when not provided by the user.
 //
-//nolint:gomnd
+//nolint:mnd
 var defaultClientConfig = ClientConfig{
 	Port:                   5033,
 	HeartbeatInterval:      time.Second * 10,
